Workflow to automatically sync with upstream.
We plan to keep this fork alive for a significant
amount of time, because it will hold an extra feature
(https://github.com/neuropoly/gitea/pull/1) that
upstream doesn't have. Automate keeping up to date
with upstream to reduce our overhead.

The scheme I chose for this is a little bit convoluted.
There are three branches involved, but this means
that the patch that we might potentially upstream
is clean.

Also, this scheme can't fix merge conflicts automatically,
but it can find and email about them automatically,
which is about as good as we can hope for.