Trigger a build when pushing a new *-git-annex tag.
We need a special token, given to us by tibdex/github-app-token,
because the default token that workflows are given is constrained:

> When you use the repository's GITHUB_TOKEN to perform tasks, events
> triggered by the GITHUB_TOKEN will not create a new workflow run.
> This prevents you from accidentally creating recursive workflow runs.
> For example, if a workflow run pushes code using the repository's
> GITHUB_TOKEN, a new workflow will not run even when the repository
> contains a workflow configured to run when push events occur.

- https://docs.github.com/en/actions/security-guides/automatic-token-authentication#using-the-github_token-in-a-workflow

See also https://github.community/t/push-from-workflow-not-doesnt-trigger-on-push-tags-workflow/149151/2

GitHub recommends https://docs.github.com/en/actions/using-workflows/reusing-workflows#calling-a-reusable-workflow
for solving this, but that seems a lot less flexible to me. I'd rather be able
to choose whether to trigger a build by pushing a tag manually or
having the bot do it.