Test uploading/downloading Private annex repos
While writing this, try rearranging the pattern so that
the Private repos are only cloned once, but are written
from multiple writers with different permissions.

This lets us test that `git annex copy`, which tries to
invoke `rsync`, is correctly blocked when it should be.
Otherwise the test would be blocked at 'git clone', which
wouldn't test the git-annex permissions at all.

But: doing this means now not testing that `git annex get`
is correctly blocked when it should be. :/

In passing, make sure to catch doGenerateRandomFile() errors.