Add a UUID to withCtxKeyFile()
This allows nesting withCtxKeyFile()s using the *same user*;
which, yes, is redundant, but might happen in a complex test
easily enough, e.g. if you're switching back and forth
between users.

TODO: a nicer solution would be to keep a cache.
That would also help with the test slowness by avoiding
generating unnecessary RSA keys.

But withKeyFile() doesn't expose the keys it generates
or let us control what they should be so we'd need to patch it first.