Rewrite filecmp to be safer.
Thanks to io.ReadFull(), this now avoids the rare case
where the read streams could get de-synced if one did a short-read
without the other, incorrectly reporting false without scanning
the complete files.

Also moves it to util (and rename it FileCmp so it can be called
from elsewhere).

Also implements checking the file size before scanning,
a feature that seems like it was once in https://stackoverflow.com/a/30038571
judging from the comments, but was since removed.

Co-authored-by: Mathieu Guay-Paquet <mathieu.guay-paquet@polymtl.ca>