git-annex: views for annex files
- show "Stored with Git Annex" on the view page
- show content's size (instead of the pointer's)
- hide annex-symlinks (LFS files never render with that quirk)
- use /media instead of the /src link for the download link (as LFS files do)

To reduce future merge conflicts, the patch tries not to restructure
the code, but instead just pastes in an extra case. But there are
obvious ways the code could be slimmed and factored; if upstream
eventually accepts git-annex as a feature, that should probably happen.

This adds tests. Especially, it adds .md, .txt, and .rst, .markdown
files to the test fixture, git-annex-smudge(1)ing the first two and
annex-symlinking the latter two, to make sure the UI detects it should
dig out and display their content, as it does for LFS-stored text files.