Allow creating/editing the _secret_ associated with each webhook from the API.
This is a signing key used by Gitea/Gogs hooks to authenticate themselves to the
receiver. Previously, this was not controllable via the API, meaning it was
impossible to set up a functioning Gitea/Gogs hook from the API, it had to be
done via the web UI instead.

This doesn't allow viewing the existing secret from the API, though it is possible for
users to find it out by inspecting the source code on the hook edit page in the web UI.