git-annex: add tests for push-to-create
This adds tests for creating a repository using the push-to-create
feature. The tests are the equivalent of doing
```
git remote add origin <url>
git annex sync --content
```
in a local git-annex repository.

The steps it does are:
1. create a local git-annex repository
2. add a non-existing repository as a remote
3. sync using `git annex sync --content`

It then checks that:
- the repository was indeed created
- the default branch matches the HEAD of the local repository
- all annexed files exist on the remote and have the correct content

This is done both for a repository in a user's namespace and for a
repository in an organization's namespace.